docker-build/response-automation/requirements.txt
flask>=3.0.3
python-dotenv>=1.0.1
requests>=2.31.0
# Docker SDK to control containers
docker>=7.0.0
# optional: interact with PostgreSQL, Prometheus, Grafana, OpenSearch
psycopg2-binary>=2.9.9
prometheus-api-client>=0.5.2
grafana-client>=3.0.0
opensearch-py>=2.4.2
# retries & logging
rich>=13.7.1
tenacity>=8.2.3
# OpenTelemetry for tracing
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0
opentelemetry-instrumentation-flask>=0.44b0
opentelemetry-instrumentation-logging>=0.44b0